Fixed a bunch of scorecard and leaderboard stuff

+++ PutPutProject/PutPutApp/views.py
@@ -4,6 +4,7 @@ from django.shortcuts import redirect, render, get_object_or_404
 from django.http import HttpResponse, HttpResponseRedirect
 from django.urls import reverse
 from datetime import date
+from django.db.models import Sum
 
 from PutPutApp.forms import *
 from .models import Drink, Orders, Profile, Score, Calendar, SponsorRequest, User
@@ -32,6 +33,8 @@ def register(request):
 
 @login_required
 def menu(request):
+    if not request.user.profile.player:
+        return HttpResponseRedirect('/dashboard')
     if request.method == "GET":
         drink_menu = Drink.objects.all()
         return render(
@@ -53,6 +56,8 @@ def menu(request):
 
 @login_required
 def orders(request):
+    if not request.user.profile.barkeep:
+        return HttpResponseRedirect('/dashboard')
     orders = Orders.objects.all()
     return render(
             request, "drinks/orders.html",
@@ -61,6 +66,8 @@ def orders(request):
 
 @login_required
 def fulfill_order(request, order_id):
+    if not request.user.profile.barkeep:
+        return HttpResponseRedirect('/dashboard')
     if request.method == "POST":
         order = get_object_or_404(Orders, pk=order_id)
         order.delete()
@@ -68,6 +75,8 @@ def fulfill_order(request, order_id):
 
 @login_required
 def sponsor_requests(request):
+    if not request.user.profile.manager:
+        return HttpResponseRedirect('/dashboard')
     if request.method == "GET":
         events = Calendar.objects.all()
         reservation_requests = SponsorRequest.objects.all()
@@ -78,6 +87,8 @@ def sponsor_requests(request):
 
 @login_required
 def approve_tournament(request, sponsor_request_id):
+    if not request.user.profile.manager:
+        return HttpResponseRedirect('/dashboard')
     if request.method == "POST":
         reservation = get_object_or_404(SponsorRequest, pk=sponsor_request_id)
         if "Approve" in request.POST:
@@ -91,6 +102,8 @@ def approve_tournament(request, sponsor_request_id):
 
 @login_required
 def sponsor(request):
+    if not request.user.profile.sponsor:
+        return HttpResponseRedirect('/dashboard')
     if request.method == "POST":
         form = SponsorForm(request.POST)
         if form.is_valid():
@@ -110,6 +123,8 @@ def sponsor(request):
 
 @login_required
 def manage_menu(request):
+    if not request.user.profile.manager:
+        return HttpResponseRedirect('/dashboard')
     if request.method == "GET":
         drink_menu = Drink.objects.all()
         return render(request, 'drinks/manage_menu.html',
@@ -139,43 +154,76 @@ def manage_menu(request):
 
 @login_required
 def scorecard(request):
+    if not request.user.profile.player:
+        return HttpResponseRedirect('/dashboard')
     if request.method == "GET":
-        scores = Score.objects.filter(user__exact=request.user).filter(day__exact=date.today())
-        return render(request, 'score/scorecard.html', {'scores' : scores, 'form': ScorecardForm})
+        scores = Score.objects.filter(user__exact=request.user).filter(day__exact=date.today()).order_by('hole').values('hole', 'num_strokes', 'par')
+        total_score = Score.objects.filter(user__exact=request.user).filter(day__exact=date.today()).aggregate(Sum('num_strokes'))['num_strokes__sum']
+        total_par = Score.objects.filter(user__exact=request.user).filter(day__exact=date.today()).aggregate(Sum('par'))['par__sum']
+        holes = []
+        for i in range(18):
+            holes.append({'hole' : i+1, 'num_strokes' : 0})
+        for i in scores:
+            holes[i['hole'] - 1] = i
+        return render(request, 'score/scorecard.html', {'scores' : holes, 'form' : ScorecardForm, 'total_score' : total_score, 'total_par' : total_par})
     if request.method == "POST":
         form = ScorecardForm(request.POST)
         if form.is_valid():
-
             score = Score(day=date.today(), hole=request.POST['hole'], num_strokes=request.POST['num_strokes'])
             score.save()
             score.user.add(request.user)
+            score.save()
     return HttpResponseRedirect(request.path_info)
             
 
 @login_required
 def leaderboard(request):
-    return render(request, "PutPutApp/leaderboard.html")
-
+    if request.method == "GET":
+        scores = Score.objects.filter(day__exact=date.today()).values('user').annotate(total_score=Sum('num_strokes'), total_par=Sum('par')).order_by('total_score').values('total_score', 'total_par', 'user__first_name', 'user__last_name')
+        for i in range(len(scores)):
+            scores[i]['count'] = i + 1
+        return render(request, 'score/leaderboard.html', {'leaderboard' : scores})
 
 @login_required
 def manage_users(request):
+    if not request.user.profile.manager:
+        return HttpResponseRedirect('/dashboard')
     if request.method == "GET":
-        users = Profile.objects.all()
-        return render(request, 'manager/manage_users.html',
-                {
-                    "curUser":request.user,
-                    "users": users,
-                    "form":ManageUserForm,
-                }
-        )
+        users = Profile.objects.all().values('id', 'user__first_name', 'user__last_name', 'player', 'barkeep', 'sponsor', 'manager')
+        print(users)
+        return render(request, 'manager/manage_users.html', {"users": users})
     if request.method == "POST":
-        form = ManageUserForm(request.POST)
-        if form.is_valid():
-            user_id = request.POST['user']
-            user_type = request.POST['user_type']
-            user = get_object_or_404(Profile, pk=user_id)
-            user.user_type = user_type[0]
-            # print(user, user.user_type)
+        new_permissions = dict(request.POST) 
+        new_permissions.pop('csrfmiddlewaretoken')
+        for key, value in new_permissions.items():
+            user = Profile.objects.get(id=key)
+            user.manager = 'manager' in value
+            user.sponsor = 'sponsor' in value
+            user.barkeep = 'barkeep' in value
+            user.player  = 'player' in value
             user.save()
-            print(user.user_type)
         return HttpResponseRedirect(request.path_info)
+        
+    
+#    if request.method == "GET":
+#        users = Profile.objects.all()
+#        return render(request, 'manager/manage_users.html',
+#                {
+#                    "curUser":request.user,
+#                    "users": users,
+#                    "form":ManageUserForm,
+#                }
+#        )
+#    if request.method == "POST":
+#        form = ManageUserForm(request.POST)
+#        if form.is_valid():
+#            user_id = request.POST['user']
+#            user_type = request.POST['user_type']
+#            user = get_object_or_404(Profile, pk=user_id)
+#            user.user_type = user_type[0]
+#            # print(user, user.user_type)
+#            user.save()
+#            print(user.user_type)
+#        return HttpResponseRedirect(request.path_info)
+
+
